Refactor(eos_designs): Move eos_designs_shared_utils to PyAVD
