+++ ansible_collections/arista/avd/plugins/plugin_utils/eos_designs_shared_utils/MOVED_TO_PYAVD.txt
@@ -1,0 +1,1 @@
+The eos_designs_facts Python module has been moved to PyAVD in the folder /python-avd/pyavd/_eos_designs/shared_utils/
