--- ansible_collections/arista/avd/plugins/plugin_utils/eos_designs_shared_utils/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (c) 2023-2024 Arista Networks, Inc.
-# Use of this source code is governed by the Apache License 2.0
-# that can be found in the LICENSE file.
-from .shared_utils import SharedUtils
-
-__all__ = ["SharedUtils"]
