+++ python-avd/pyavd/_eos_designs/shared_utils/connected_endpoints_keys.py
@@ -6,11 +6,11 @@ from __future__ import annotations
 from functools import cached_property
 from typing import TYPE_CHECKING
 
-from ansible_collections.arista.avd.plugins.filter.convert_dicts import convert_dicts
-from ansible_collections.arista.avd.plugins.plugin_utils.utils import get
+from ...j2filters.convert_dicts import convert_dicts
+from ...vendor.utils import get
 
 if TYPE_CHECKING:
-    from .shared_utils import SharedUtils
+    from . import SharedUtils
 
 # NOTE: there is a static list of default endpoint keys in the fabric connected endpoints documentation templates.
 DEFAULT_CONNECTED_ENDPOINTS_KEYS = [
