+++ python-avd/pyavd/_eos_designs/shared_utils/cv_topology.py
@@ -6,11 +6,11 @@ from __future__ import annotations
 from functools import cached_property
 from typing import TYPE_CHECKING
 
-from ansible_collections.arista.avd.plugins.filter.range_expand import range_expand
-from ansible_collections.arista.avd.plugins.plugin_utils.utils import get, get_item
+from ...vendor.j2.filter.range_expand import range_expand
+from ...vendor.utils import get, get_item
 
 if TYPE_CHECKING:
-    from .shared_utils import SharedUtils
+    from . import SharedUtils
 
 
 class CvTopology:
