+++ python-avd/pyavd/_eos_designs/shared_utils/inband_management.py
@@ -7,11 +7,11 @@ from functools import cached_property
 from ipaddress import ip_network
 from typing import TYPE_CHECKING
 
-from ansible_collections.arista.avd.plugins.plugin_utils.errors import AristaAvdError, AristaAvdMissingVariableError
-from ansible_collections.arista.avd.plugins.plugin_utils.utils import default, get
+from ...vendor.errors import AristaAvdError, AristaAvdMissingVariableError
+from ...vendor.utils import default, get
 
 if TYPE_CHECKING:
-    from .shared_utils import SharedUtils
+    from . import SharedUtils
 
 
 class InbandManagementMixin:
