+++ python-avd/pyavd/_eos_designs/shared_utils/l3_interfaces.py
@@ -6,11 +6,11 @@ from __future__ import annotations
 from functools import cached_property
 from typing import TYPE_CHECKING
 
-from ansible_collections.arista.avd.plugins.plugin_utils.merge import merge
-from ansible_collections.arista.avd.plugins.plugin_utils.utils import get, get_item
+from ...vendor.merge import merge
+from ...vendor.utils import get, get_item
 
 if TYPE_CHECKING:
-    from .shared_utils import SharedUtils
+    from . import SharedUtils
 
 
 class L3InterfacesMixin:
