+++ python-avd/pyavd/get_avd_facts.py
@@ -1,9 +1,12 @@
 # Copyright (c) 2023-2024 Arista Networks, Inc.
 # Use of this source code is governed by the Apache License 2.0
 # that can be found in the LICENSE file.
+from __future__ import annotations
+
 from collections import ChainMap
 
 from ._eos_designs.eos_designs_facts import EosDesignsFacts
+from ._eos_designs.shared_utils import SharedUtils
 
 
 def get_avd_facts(all_inputs: dict[str, dict]) -> dict[str, dict]:
@@ -70,11 +73,6 @@ def _create_avd_switch_facts_instances(all_inputs: dict[str, dict]) -> dict:
             }
             ```
     """
-    # pylint: disable=import-outside-toplevel
-    from .vendor.eos_designs.eos_designs_shared_utils import SharedUtils
-
-    # pylint: enable=import-outside-toplevel
-
     avd_switch_facts = {}
     for hostname, hostvars in all_inputs.items():
         # Set 'inventory_hostname' on the input variables, to keep compatibility with Ansible focused code.
