+++ python-avd/pyavd/validate_inputs.py
@@ -1,6 +1,9 @@
 # Copyright (c) 2023-2024 Arista Networks, Inc.
 # Use of this source code is governed by the Apache License 2.0
 # that can be found in the LICENSE file.
+from __future__ import annotations
+
+from ._eos_designs.shared_utils import SharedUtils
 from .avd_schema_tools import AvdSchemaTools
 from .constants import EOS_DESIGNS_SCHEMA_ID
 from .validation_result import ValidationResult
@@ -20,8 +23,6 @@ def validate_inputs(inputs: dict) -> ValidationResult:
     Returns:
         Validation result object with any validation errors or deprecation warnings.
     """
-    from .vendor.eos_designs.eos_designs_shared_utils import SharedUtils  # pylint: disable=import-outside-toplevel
-
     # Initialize a global instance of eos_designs_schema_tools
     global eos_designs_schema_tools
     if eos_designs_schema_tools is None:
